requirements.txt
streamlit
arxiv
google-generativeai
PyMuPDF
faiss-cpu
requests
sentence-transformers
fpdf2